Added empty camera object
This is the camera that will follow the player for most of the game. I
only added an empty sprite (origin is center 16,16) and added an empty
obj_camera. This does not yet have any events, code, no instance in the
room, and is not set up as a view yet. Work needs to be done, refer to
RPG tutorial for reference